Get default options for bias corrections solver

Get default options for bias corrections solver

Usage

downscale_control(
  solve_fun = "solve_biascorr",
  algorithm = "NLOPT_LN_SBPLX",
  xtol_rel = 1e-20,
  xtol_abs = 1e-20,
  maxeval = 1600,
  MAX_EXP = log(.Machine$double.xmax),
  cutoff = 0,
  redo = 2,
  max_diff = 10^-8,
  err.txt = "",
  prior_weights = 0
)

Arguments

solve_fun

Downscaling algorithm (should be solve_biascorr.mnl)).

algorithm

Solver algorithm (see the nloptr package for documentation and more detail).

xtol_rel

Relative tolerance of solver.

xtol_abs

Absolute solver tolerance.

maxeval

Maximum evaluation of solver.

MAX_EXP

Numerical cutoff for MNL function.

cutoff

Optional cutoff to avoid MNL values close to zero.

redo

Maximum number of repeats

max_diff

If difference to targets is larger, redo the estimation (helps to avoid convergence errors)

err.txt

Error text for caller identification (used for debugging)

prior_weights

Scalar between 0 and 1; if both betas and priors are specified this value gives the weight of the priors.

Value

List with default options for bias correction solver

Details

Call this function if you want to change default options for the bias corrections solver.

Examples

opts1 = downscale_control()