Bayesian logit model with Pólya Gamma prior with MCMC

Bayesian logit model with Pólya Gamma prior with MCMC

Usage

mnlogit(X, Y, baseline = ncol(Y), niter = 1000, nburn = 500, A0 = 10^4)

Arguments

X

An n by k matrix of explanatory variables

Y

An n by p matrix of dependent variables

baseline

Baseline class for estimation. Parameters will be set to zero. Defaults to the p-th column.

niter

Total number of MCMC draws

nburn

Burn-in draws for MCMC

A0

Prior variance scalar for all slope coefficients

Value

A list containing

postb A k x p x (niter - nburn dimensions) array containing posterior draws of the slope coefficients.

marginal_fx A k x p x (niter - nburn dimensions) array containing posterior draws of marginal effects.

X, Y, baseline The matrices of explanatory and dependent variables, as defined above and the baseline class.

Details

MCMC estimation of a multinomial logit model following Polson et al. (2013).

References

Nicholas G. Polson, James G. Scott, and Jesse Windle. Bayesian inference for logistic models using Polya-Gamma latent variables. Journal of the American statistical Association 108.504 (2013): 1339-1349.

Examples

n <- 100
p <- 3
k <- 2
X <- cbind(1, matrix(rnorm(n * (k - 1), 0, 2), n, k - 1))
BETA <- matrix(sample(c(-3:3), k * p, replace = TRUE), k, p)
BETA[, p] <- 0
Y <- exp(X %*% BETA) / rowSums(exp(X %*% BETA))
res1 <- mnlogit(X, Y)
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |==                                                                    |   4%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |=======                                                               |  11%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |=========                                                             |  14%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  16%  |                                                                              |============                                                          |  17%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |================                                                      |  24%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |=======================                                               |  34%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |============================                                          |  41%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |==============================                                        |  44%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |===================================                                   |  51%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  54%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |============================================                          |  64%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |===================================================                   |  74%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  79%  |                                                                              |========================================================              |  80%  |                                                                              |========================================================              |  81%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |==========================================================            |  84%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |===============================================================       |  91%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |=================================================================     |  94%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================|  99%  |                                                                              |======================================================================| 100%
print(BETA)
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]   -1    1    0
print(apply(res1$postb, c(1, 2), mean))
#>             [,1]        [,2] [,3]
#> [1,]  0.01199738 -0.03041672    0
#> [2,] -1.05266934  1.09022027    0

require(tidyr)
#> Loading required package: tidyr
Y = dplyr::filter(argentina_luc,lu.from == "Cropland" & Ts == 2000) %>%
   pivot_wider(names_from = lu.to)
X = argentina_df$xmat %>% tidyr::pivot_wider(names_from = "ks") %>%
   dplyr::arrange(match(ns,Y$ns))
Y = Y %>% dplyr::select(-c(lu.from,Ts,ns))
X = X %>% dplyr::select(-c(ns))
res1 <- mnlogit(as.matrix(X), as.matrix(Y),baseline = which(colnames(Y) == "Cropland"),
          niter = 100,nburn = 50)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%
print(apply(res1$postb, c(1, 2), mean))
#>                  Cropland        Forest     OtherLand       Pasture
#> Intercept               0 -4.529333e+00 -7.085907e+00 -7.503696e+00
#> Cropland                0 -4.794961e-04  1.307174e-02 -7.453563e-03
#> Forest                  0 -1.316549e-02  4.984865e-03  1.054297e-02
#> OtherLand               0 -1.388280e-02  1.586195e-02  1.314232e-02
#> Pasture                 0  4.392545e-03 -9.450790e-05 -2.074675e-02
#> Plantations             0 -7.174321e-01 -8.483023e-02 -1.351351e-02
#> Urban                   0  4.857346e-01 -9.780817e-01 -1.043831e+00
#> MeanTemp                0  4.640202e-02  1.257020e-03 -2.518854e-02
#> MeanPrecip              0 -3.644472e-04 -2.532720e-03 -2.481193e-04
#> HarvWood                0  9.025559e-01  9.076491e-01  1.142085e+00
#> HarvCost                0 -5.361520e-03 -6.859264e-03 -3.786091e-03
#> GRASYLD                 0 -1.704676e-01 -1.754392e-01 -1.713227e-01
#> meanTimeToMarket        0 -1.474450e-03 -5.089997e-04  3.447891e-04
#> totPop                  0 -9.148209e-04 -1.963073e-04 -1.053115e-03
#> ruralPop                0 -1.589169e-01  1.073801e+00  7.842843e-02
#> Altitude                0 -5.733268e-04  8.660088e-04  8.563083e-04
#> Slope                   0 -2.272171e-02 -4.998775e-03 -1.265940e-02
#> Soil2_Medium            0 -1.227044e+00 -2.210353e-01  8.183295e-01
#> Soil3_Heavy             0 -2.505465e+00  1.809708e+00  3.629196e+00
#> Soil4_Stony             0 -6.582842e-01 -2.453459e-01  4.298497e-01
#> MEANYLD                 0 -2.968527e+01  1.883433e+01 -4.098025e+01
#> HI_MEAN                 0  9.903289e+00  5.725272e+00  2.402574e+00
#> LI_MEAN                 0  4.739950e+00 -1.685170e+01  2.045177e+01
#> IR_MEAN                 0  6.627606e+00 -5.328829e+00  9.684145e+00
#> SS_MEAN                 0  9.123539e+00 -2.347392e+00  8.662462e+00
#>                    Plantations         Urban
#> Intercept        -4.9924170934 -4.870836e+00
#> Cropland         -0.0017332813 -3.235512e-03
#> Forest           -0.0057183912 -7.173270e-03
#> OtherLand        -0.0038339055  8.487686e-03
#> Pasture          -0.0070140494 -1.957869e-02
#> Plantations      -0.0044951081 -7.517060e-01
#> Urban             0.1912258796 -1.300845e-01
#> MeanTemp          0.0165732195  8.988940e-03
#> MeanPrecip       -0.0003281360  4.233999e-04
#> HarvWood          1.0397862948  2.134040e-01
#> HarvCost         -0.0055146004 -5.839288e-03
#> GRASYLD           0.0080880510 -4.082310e-02
#> meanTimeToMarket  0.0007655606 -6.595466e-04
#> totPop           -0.0004823327 -8.405439e-04
#> ruralPop         -0.8704598626 -1.885395e+00
#> Altitude         -0.0002580388 -4.517942e-04
#> Slope            -0.0237438326 -4.589672e-03
#> Soil2_Medium      0.0962293253  5.259809e-01
#> Soil3_Heavy       1.9607132773 -7.837969e-01
#> Soil4_Stony       0.0393099811 -1.940071e+00
#> MEANYLD           0.3796221322 -3.611587e+01
#> HI_MEAN           6.2431310340  1.432740e+01
#> LI_MEAN          -6.1494844894  1.591078e+00
#> IR_MEAN          -0.8953474998  8.726106e+00
#> SS_MEAN           0.4177418794  1.172179e+01