Bayesian logit model with Pólya Gamma prior with MCMC

Bayesian logit model with Pólya Gamma prior with MCMC

Usage

mnlogit(X, Y, baseline = ncol(Y), niter = 1000, nburn = 500, A0 = 10^4)

Arguments

X

An n by k matrix of explanatory variables

Y

An n by p matrix of dependent variables

baseline

Baseline class for estimation. Parameters will be set to zero. Defaults to the p-th column.

niter

Total number of MCMC draws

nburn

Burn-in draws for MCMC

A0

Prior variance scalar for all slope coefficients

Value

A list containing

postb A k x p x (niter - nburn dimensions) array containing posterior draws of the slope coefficients.

marginal_fx A k x p x (niter - nburn dimensions) array containing posterior draws of marginal effects.

X, Y, baseline The matrices of explanatory and dependent variables, as defined above and the baseline class.

Details

MCMC estimation of a multinomial logit model following Polson et al. (2013).

References

Nicholas G. Polson, James G. Scott, and Jesse Windle. Bayesian inference for logistic models using Polya-Gamma latent variables. Journal of the American statistical Association 108.504 (2013): 1339-1349.

Examples

n <- 100
p <- 3
k <- 2
X <- cbind(1, matrix(rnorm(n * (k - 1), 0, 2), n, k - 1))
BETA <- matrix(sample(c(-3:3), k * p, replace = TRUE), k, p)
BETA[, p] <- 0
Y <- exp(X %*% BETA) / rowSums(exp(X %*% BETA))
res1 <- mnlogit(X, Y)
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |==                                                                    |   4%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |=======                                                               |  11%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |=========                                                             |  14%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  16%  |                                                                              |============                                                          |  17%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |================                                                      |  24%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |=======================                                               |  34%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |============================                                          |  41%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |==============================                                        |  44%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |===================================                                   |  51%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  54%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |============================================                          |  64%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |===================================================                   |  74%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  79%  |                                                                              |========================================================              |  80%  |                                                                              |========================================================              |  81%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |==========================================================            |  84%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |===============================================================       |  91%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |=================================================================     |  94%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================|  99%  |                                                                              |======================================================================| 100%
print(BETA)
#>      [,1] [,2] [,3]
#> [1,]    1    2    0
#> [2,]   -2    0    0
print(apply(res1$postb, c(1, 2), mean))
#>           [,1]       [,2] [,3]
#> [1,]  1.074283  2.1627998    0
#> [2,] -2.182640 -0.0213314    0

require(tidyr)
Y = dplyr::filter(argentina_luc,lu.from == "Cropland" & Ts == 2000) %>%
   pivot_wider(names_from = lu.to)
X = argentina_df$xmat %>% tidyr::pivot_wider(names_from = "ks") %>%
   dplyr::arrange(match(ns,Y$ns))
Y = Y %>% dplyr::select(-c(lu.from,Ts,ns))
X = X %>% dplyr::select(-c(ns))
res1 <- mnlogit(as.matrix(X), as.matrix(Y),baseline = which(colnames(Y) == "Cropland"),
          niter = 100,nburn = 50)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%
print(apply(res1$postb, c(1, 2), mean))
#>                  Cropland        Forest     OtherLand       Pasture
#> Intercept               0 -5.289579e+00 -3.1292984593 -8.725199e+00
#> Cropland                0  1.464205e-03 -0.0115843412 -1.319992e-02
#> Forest                  0 -2.315198e-02 -0.0042704336  4.334065e-03
#> OtherLand               0  2.048044e-02  0.0115262447  7.878238e-03
#> Pasture                 0 -2.167469e-02 -0.0037094959 -1.105908e-03
#> Plantations             0 -1.195511e-02 -0.3057127323  1.572946e-02
#> Urban                   0 -7.681449e-01 -0.5017972128 -4.789008e-02
#> MeanTemp                0  8.582911e-03 -0.0086219592 -2.042669e-02
#> MeanPrecip              0  7.462977e-04  0.0001412098 -6.901696e-04
#> HarvWood                0 -9.349695e-01  0.0266147878  8.962797e-01
#> HarvCost                0 -1.857787e-04 -0.0039130855 -5.066799e-04
#> GRASYLD                 0  1.402563e-01 -0.2700196095 -1.220189e-02
#> meanTimeToMarket        0 -8.710591e-04 -0.0029298558  2.514899e-04
#> totPop                  0 -5.779385e-05 -0.0001714562 -5.261873e-04
#> ruralPop                0 -1.706042e+00 -0.2397548069  2.087703e-01
#> Altitude                0 -5.138179e-05 -0.0012640558 -1.233918e-04
#> Slope                   0  2.613164e-02 -0.0580291594  2.916201e-02
#> Soil2_Medium            0 -1.469595e+00 -2.1438781142  1.926695e+00
#> Soil3_Heavy             0 -1.972856e+00 -0.7538888501  3.181945e+00
#> Soil4_Stony             0 -1.631885e+00 -1.4371918452  1.410379e+00
#> MEANYLD                 0  1.025548e+01 -4.4852823335 -2.395510e+01
#> HI_MEAN                 0 -9.589345e+00  7.0547552229  1.021390e+01
#> LI_MEAN                 0  5.139382e+00 -8.7483953581  7.186502e-01
#> IR_MEAN                 0 -3.006234e+00  1.0570878545  5.841392e+00
#> SS_MEAN                 0 -2.434316e+00  5.7512446052  7.276656e+00
#>                    Plantations         Urban
#> Intercept        -8.5329123599 -6.635683e+00
#> Cropland          0.0002397458  2.180647e-02
#> Forest           -0.0118458637  1.734147e-02
#> OtherLand         0.0035671996 -1.680285e-02
#> Pasture          -0.0117637072  3.567249e-03
#> Plantations       0.0038624030 -1.973654e-01
#> Urban            -0.1232417843  1.853840e-01
#> MeanTemp          0.0277748057  3.866961e-02
#> MeanPrecip       -0.0003278452  1.926074e-04
#> HarvWood          0.8063647842  5.626598e-01
#> HarvCost         -0.0099011129  6.499833e-03
#> GRASYLD           0.0578377210 -7.865317e-02
#> meanTimeToMarket  0.0005616744  3.602038e-04
#> totPop           -0.0002042795 -6.871762e-04
#> ruralPop         -1.3265860185  1.149543e-01
#> Altitude         -0.0001861130 -7.742193e-04
#> Slope            -0.0184286680 -4.471870e-02
#> Soil2_Medium      3.8119065884 -6.537951e-01
#> Soil3_Heavy       5.3330820283 -8.945976e-01
#> Soil4_Stony       3.8759074505 -5.908077e-01
#> MEANYLD          10.2442906048  1.138875e+01
#> HI_MEAN           2.9413551461  5.737642e+00
#> LI_MEAN          -7.0023507922 -1.196511e+01
#> IR_MEAN          -3.6973832935 -4.287661e+00
#> SS_MEAN          -2.4212497121 -6.674519e-01