Bayesian logit model with Pólya Gamma prior with MCMC

Bayesian logit model with Pólya Gamma prior with MCMC

Usage

mnlogit(X, Y, baseline = ncol(Y), niter = 1000, nburn = 500, A0 = 10^4)

Arguments

X

An n by k matrix of explanatory variables

Y

An n by p matrix of dependent variables

baseline

Baseline class for estimation. Parameters will be set to zero. Defaults to the p-th column.

niter

Total number of MCMC draws

nburn

Burn-in draws for MCMC

A0

Prior variance scalar for all slope coefficients

Value

A list containing

postb A k x p x (niter - nburn dimensions) array containing posterior draws of the slope coefficients.

marginal_fx A k x p x (niter - nburn dimensions) array containing posterior draws of marginal effects.

X, Y, baseline The matrices of explanatory and dependent variables, as defined above and the baseline class.

Details

MCMC estimation of a multinomial logit model following Polson et al. (2013).

References

Nicholas G. Polson, James G. Scott, and Jesse Windle. Bayesian inference for logistic models using Polya-Gamma latent variables. Journal of the American statistical Association 108.504 (2013): 1339-1349.

Examples

n <- 100
p <- 3
k <- 2
X <- cbind(1, matrix(rnorm(n * (k - 1), 0, 2), n, k - 1))
BETA <- matrix(sample(c(-3:3), k * p, replace = TRUE), k, p)
BETA[, p] <- 0
Y <- exp(X %*% BETA) / rowSums(exp(X %*% BETA))
res1 <- mnlogit(X, Y)
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |==                                                                    |   4%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |=======                                                               |  11%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |=========                                                             |  14%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  16%  |                                                                              |============                                                          |  17%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |================                                                      |  24%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |=======================                                               |  34%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |============================                                          |  41%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |==============================                                        |  44%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |===================================                                   |  51%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  54%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |============================================                          |  64%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |===================================================                   |  74%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  79%  |                                                                              |========================================================              |  80%  |                                                                              |========================================================              |  81%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |==========================================================            |  84%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |===============================================================       |  91%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |=================================================================     |  94%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================|  99%  |                                                                              |======================================================================| 100%
print(BETA)
#>      [,1] [,2] [,3]
#> [1,]   -1   -3    0
#> [2,]   -3   -2    0
print(apply(res1$postb, c(1, 2), mean))
#>           [,1]      [,2] [,3]
#> [1,] -1.047770 -3.351063    0
#> [2,] -3.070239 -1.743306    0

require(tidyr)
#> Loading required package: tidyr
#> Warning: package ‘tidyr’ was built under R version 4.1.2
Y = dplyr::filter(argentina_luc,lu.from == "Cropland" & Ts == 2000) %>%
   pivot_wider(names_from = lu.to)
X = argentina_df$xmat %>% tidyr::pivot_wider(names_from = "ks") %>%
   dplyr::arrange(match(ns,Y$ns))
Y = Y %>% dplyr::select(-c(lu.from,Ts,ns))
X = X %>% dplyr::select(-c(ns))
res1 <- mnlogit(as.matrix(X), as.matrix(Y),baseline = which(colnames(Y) == "Cropland"),
          niter = 100,nburn = 50)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%
print(apply(res1$postb, c(1, 2), mean))
#>                  Cropland        Forest     OtherLand       Pasture
#> Intercept               0 -5.903810e+00 -5.1332954783 -7.6697782895
#> Cropland                0 -1.752316e-02 -0.0175236627 -0.0103150721
#> Forest                  0 -3.675627e-03 -0.0060640651 -0.0286329013
#> OtherLand               0 -3.711341e-03  0.0045825202  0.0130595936
#> Pasture                 0  6.485055e-05 -0.0142402845  0.0006306296
#> Plantations             0 -4.648604e-01 -0.3604304469  0.0281967960
#> Urban                   0 -1.237911e+00 -0.2704556473 -3.2622606015
#> MeanTemp                0 -1.590495e-02 -0.0169480945  0.0429477646
#> MeanPrecip              0  6.539379e-05 -0.0003082205 -0.0002315560
#> HarvWood                0  3.178816e-02  0.2339763915  1.3945234739
#> HarvCost                0  1.959301e-03 -0.0018930908 -0.0079705578
#> GRASYLD                 0 -5.021173e-02 -0.0548350442 -0.0292511742
#> meanTimeToMarket        0  3.814257e-04  0.0001018913  0.0010435954
#> totPop                  0 -8.238142e-04 -0.0008186776 -0.0001250847
#> ruralPop                0 -1.638575e-01 -1.5468393629 -0.4417394013
#> Altitude                0  2.642590e-04 -0.0011894816  0.0005282528
#> Slope                   0 -1.530133e-01  0.0534407772 -0.0312505831
#> Soil2_Medium            0 -9.282740e-02  1.3076128277  0.2914486687
#> Soil3_Heavy             0  7.560271e-01  2.0372638574  1.7959772398
#> Soil4_Stony             0  1.718475e-01  0.4096068671 -0.2964687288
#> MEANYLD                 0 -5.111465e+01 -6.9637824150  9.4491693707
#> HI_MEAN                 0  1.421853e+01 -6.7008505381 -1.5255689405
#> LI_MEAN                 0  1.025124e+01 10.9187336454 -3.7650661142
#> IR_MEAN                 0  1.287526e+01  1.5317470237 -2.6766029738
#> SS_MEAN                 0  1.388184e+01  1.4873810625 -1.2330139457
#>                    Plantations         Urban
#> Intercept        -8.652939e+00 -9.546786e+00
#> Cropland         -1.702905e-02  2.228774e-02
#> Forest           -2.910605e-05 -5.041216e-03
#> OtherLand         6.653189e-03  7.556376e-04
#> Pasture          -4.811344e-03 -5.560218e-03
#> Plantations       1.166669e-02 -7.072886e-01
#> Urban            -7.282405e-01  2.699558e-01
#> MeanTemp         -7.338759e-03  1.115932e-02
#> MeanPrecip       -3.818152e-04 -6.571077e-07
#> HarvWood          5.841356e-01 -9.859151e-01
#> HarvCost         -6.678927e-03  1.892411e-03
#> GRASYLD           5.947264e-02 -3.564067e-03
#> meanTimeToMarket  6.510955e-04  2.017721e-03
#> totPop           -1.472065e-04 -1.436379e-04
#> ruralPop         -1.400699e+00 -1.661817e+00
#> Altitude         -1.994883e-04  9.021366e-04
#> Slope             3.379831e-02  1.822879e-03
#> Soil2_Medium      4.020872e+00  1.906970e+00
#> Soil3_Heavy       5.802181e+00  4.801459e-01
#> Soil4_Stony       3.479486e+00  2.187770e+00
#> MEANYLD          -1.452197e+01 -2.980956e+01
#> HI_MEAN           6.523837e+00  4.756297e-01
#> LI_MEAN           3.025750e+00  1.772059e+01
#> IR_MEAN           2.220310e+00  6.655612e+00
#> SS_MEAN           2.803958e+00  5.316165e+00